Reworked the issues on Hex4D, needs testing.
