+++ Example Code/Program 1, Hello Branch and Math.txt
@@ -1,6 +1,8 @@
 @ This is just be a simple 2+2=4 program using branch to create a variable space at the start of the program.
-br m02
+br start:
 #0002
+
+start:
 ld m01 r0
 ld m01 r1
 add r0 r1 r2
